refactor: change folders name
