+++ src/app/models/city-summary.ts
