+++ src/app/models/search-result.ts
